refactor: move MergeContainersSideBySide to ModificationsUtil
