+++ Editor/Processors/TraceAndOptimize/AnimatorParser.cs
@@ -294,7 +294,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
         /// </summary>
         private IModificationsContainer FallbackParseAnimatorController(GameObject root, RuntimeAnimatorController controller)
         {
-            return MergeContainersSideBySide(controller.animationClips.Select(clip => GetParsedAnimation(root, clip)));
+            return ModificationsUtils.MergeContainersSideBySide(controller.animationClips.Select(clip => GetParsedAnimation(root, clip)));
         }
 
         private IModificationsContainer AdvancedParseAnimatorController(GameObject root, AnimatorController controller,
@@ -354,8 +354,8 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
                 AnimationClip MapClip(AnimationClip clip) => mapping.TryGetValue(clip, out var newClip) ? newClip : clip;
 
                 mergedController.MergeAsNewLayer(
-                    MergeContainersSideBySide(animationClips.Select(x => GetParsedAnimation(root, MapClip(x)))),
-                    alwaysAppliedLayer);
+                    ModificationsUtils.MergeContainersSideBySide(animationClips.Select(x =>
+                        GetParsedAnimation(root, MapClip(x)))), alwaysAppliedLayer);
             }
 
             return mergedController;
@@ -374,26 +374,6 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
 
                 foreach (var childStateMachine in stateMachine.stateMachines)
                     queue.Enqueue(childStateMachine.stateMachine);
-            }
-        }
-
-        private IModificationsContainer MergeContainersSideBySide<T>([ItemNotNull] IEnumerable<T> enumerable)
-            where T : IModificationsContainer
-        {
-            using (var enumerator = enumerable.GetEnumerator())
-            {
-                if (!enumerator.MoveNext()) return ImmutableModificationsContainer.Empty;
-                var first = enumerator.Current;
-                if (!enumerator.MoveNext()) return first;
-
-                // ReSharper disable once PossibleNullReferenceException // miss detections
-
-                // merge all properties
-                var merged = first.ToMutable();
-                do merged.MergeAsSide(enumerator.Current);
-                while (enumerator.MoveNext());
-
-                return merged;
             }
         }
